Texture system overhaul, misc sprite/font changes (...) Added TexturesystemClass in new source file DX8Texture.cpp that abstracts texture handling for all platforms & configurations. Textures are now shared between sprites that use the same image file, instead of being loaded multiple times. Texture-related functions now take C++ strings as parameters instead of C strings. TexturesystemClass can now load raw or compressed image files from disk that have been arbitrarily resized or expanded to NPOT dimensions. (New scalefactors.txt filein a texture's directory is read to accomplish this, which is written by the new texture-processing utility I've written). DirectGraphicsSprite class now uses itsXTexScale and itsYTexScale, to compute texture coordinates. DirectGraphics.SetTexture() is now used for all platforms (DirectX or SDL) to change the current texture. Sprites use itsTexIdx data member to interact with new texturing system, instead of itsTexture (has been removed). Moved all sprites declared in global scope to instead be members of classes of their category, so we can control when and where they get destructed. Projectiles.cpp and Partikelsystem.cpp were using AnimEnde variable such that it was drawing one frame too few in animated sprites. Projectiles.cpp and Partikelsystem.cpp now use itsPreCalcedRects instead of computing RECTS afresh needlessly. Projectiles.cpp and Partikelsystem.cpp now use a value for AnimEnde that corresponds to the sprites' true last frame of animation and thereby stay within itsPreCalcedRects boundaries. Menufont.png has been expanded in size and the three missing Swedish glyphs have been added. Now the Swedish translation works correctly. Font glyph widths have been hard-coded in DX8Font.cpp, so now fonts can be compressed and their image files no longer must be read directly. Moved LoadingBar and LoadingScreen sprites from Main.cpp globals to MenuClass members. EXPLOSION_MEDIUM3 particle now loads the same texture image file as EXPLOSION_MEDIUM, as they were 100% identical (saves VRAM). Corrected dimensions and yfc parameter (4 instead of 5) in Partikelsystem.cpp for EXPLOSION_BIG.

+++ Hurrican/src/SDLPort/SDL_port.h
@@ -60,13 +60,13 @@ typedef uint32_t    LPDIRECTINPUT8;
 #ifndef __WIN32__
 typedef uint32_t HINSTANCE, HANDLE, HWND, HRESULT;
 
-//DKS - Changed the members from LONG to int16_t:
+//DKS - Converted to int32_t from LONG:
 typedef struct tagRECT
 {
-    int16_t left;
-    int16_t top;
-    int16_t right;
-    int16_t bottom;
+    int32_t left;
+    int32_t top;
+    int32_t right;
+    int32_t bottom;
 } RECT;
 #endif
 typedef uint32_t LPDIRECT3D8, LPDIRECT3D9, LPDIRECT3DDEVICE8, LPDIRECT3DDEVICE9, LPDIRECTSOUND8, D3DCOLOR;
@@ -75,7 +75,7 @@ typedef uint32_t LPDIRECT3D8, LPDIRECT3D9, LPDIRECT3DDEVICE8, LPDIRECT3DDEVICE9,
 #define MIN(a,b)    ((a)<(b)?(a):(b))
 #define LIM(a,b,c)  MAX(MIN(a,c),b)
 
-#define D3DX_PI M_PI
+#define D3DX_PI ((float)(M_PI))
 
 #if SDL_BYTEORDER == SDL_LIL_ENDIAN
 #define D3DCOLOR_RGBA(r,g,b,a)  (((a)<<24) + ((r)<<16) + ((g)<<8) + (b))
